build(docs): update dependencies
