+++ packages/docs-next/.eslintrc.cjs
@@ -11,6 +11,7 @@ module.exports = {
   ],
   plugins: ["prettier"],
   rules: {
+    "@typescript-eslint/no-explicit-any": ["warn"],
     "vue/padding-line-between-blocks": ["error", "always"],
     "vue/multi-word-component-names": ["off"],
     "comma-dangle": ["error", "always-multiline"],
