+++ packages/docs-next/.vitepress/theme/styles/vitepress.scss
@@ -2,6 +2,11 @@
 
 :root {
     --vp-layout-max-width: 1550px;
+
+    --vp-c-brand-1: var(--vp-c-green-1);
+    --vp-c-brand-2: var(--vp-c-green-2);
+    --vp-c-brand-3: var(--vp-c-green-3);
+
 }
 
 .VPNavBarTitle a {
