+++ packages/docs-next/package.json
@@ -5,6 +5,7 @@
   "description": "UI components for Vue.js and CSS framework agnostic",
   "author": "Walter Tommasi <tommsi20@gmail.com>",
   "license": "MIT",
+  "type": "module",
   "scripts": {
     "docs:dev": "concurrently \"vue-docgen --watch\" \"vitepress dev\"",
     "docs:gen": "vue-docgen",
@@ -13,33 +14,33 @@
     "update": "ncu -u"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^6.4.0",
-    "@fortawesome/free-solid-svg-icons": "^6.4.0",
-    "@fortawesome/vue-fontawesome": "^3.0.3",
+    "@fortawesome/fontawesome-svg-core": "^6.5.1",
+    "@fortawesome/free-solid-svg-icons": "^6.5.1",
+    "@fortawesome/vue-fontawesome": "^3.0.5",
     "@highlightjs/vue-plugin": "2.1.2",
     "@oruga-ui/theme-bootstrap": "^0.4.0",
     "@oruga-ui/theme-bulma": "^0.2.11",
     "@oruga-ui/theme-oruga": "^0.0.1",
-    "highlight.js": "11.8.0",
-    "markdown-it": "^13.0.1",
+    "highlight.js": "11.9.0",
+    "markdown-it": "^13.0.2",
     "markdown-it-highlightjs": "^4.0.1"
   },
   "devDependencies": {
-    "@types/node": "^18.16.19",
-    "@vue/eslint-config-prettier": "^7.1.0",
-    "@vue/eslint-config-typescript": "^11.0.3",
+    "@types/node": "^20.10.3",
+    "@vue/eslint-config-prettier": "^8.0.0",
+    "@vue/eslint-config-typescript": "^12.0.0",
     "@vue/tsconfig": "^0.4.0",
-    "concurrently": "8.2.0",
-    "eslint": "^8.45.0",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-vue": "^9.15.1",
-    "npm-check-updates": "^16.10.16",
-    "prettier": "^2.8.8",
-    "sass": "^1.63.6",
-    "typescript": "5.1.6",
-    "vitepress": "^1.0.0-beta.5",
-    "vue": "^3.3.4",
+    "concurrently": "8.2.2",
+    "eslint": "^8.55.0",
+    "eslint-plugin-prettier": "^5.0.1",
+    "eslint-plugin-vue": "^9.19.2",
+    "npm-check-updates": "^16.14.11",
+    "prettier": "^3.1.0",
+    "sass": "^1.69.5",
+    "typescript": "5.3.2",
+    "vitepress": "^1.0.0-rc.31",
+    "vue": "^3.3.10",
     "vue-docgen-cli": "4.67.0",
-    "vue-tsc": "^1.8.5"
+    "vue-tsc": "^1.8.25"
   }
 }
