+++ packages/docs-next/tsconfig.json
@@ -7,7 +7,13 @@
     "resolveJsonModule": true,
     "allowJs": true,
     "baseUrl": ".",
-    "types": ["vite/client"],
+    "types": [ "node", "vite/client"],
   },
-  "include": [".vitepress/**/*"]
+  "include": [
+    ".vitepress/**/*",
+    "docgen.config.ts"
+  ],
+  "exclude": [
+    "node_modules"
+  ]
 }
